Removed the "string hack"
Runtime attribute disk_size requires type int in its string. Previously we kind of cheated by using strings to coerce a float into an int. With the built-in ceil() we can now make things a lot simplier. Ceil() will of course round up, but that was the end result of the string hack anyway.